+++ ui/src/windows/editor/core/workspace/editor-workspace.service.ts
