+++ ui/src/windows/editor/editor-select-directory/editor-select-directory.component.ts
@@ -3,7 +3,7 @@ import { MatButtonModule } from "@angular/material/button";
 import { getElectronApi } from "../../../shared/electron";
 import { EditorStateService } from "../core/state/editor-state.service";
 import { EditorSidebarPaneService } from "../core/panes/editor-sidebar-pane.service";
-import { EditorWorkspaceService } from "../core/services/editor-workspace.service";
+import { EditorWorkspaceService } from "../core/workspace/editor-workspace.service";
 
 /**
  * Shown when there isnt any select directory and allows users to select a directory
